move test scaffold to common
